users/requirements.txt
Django~=5.0
requests
djangorestframework~=3.0
psycopg2-binary~=2.0
djangorestframework_simplejwt
django-environ
bcrypt
django-cors-headers
